updating the script to better separation
